PR comments resolve
